+++ src/activityLog/ActivityLogTable.jsx
@@ -953,7 +953,7 @@ export function ActivityLogTable() {
                       <Dropdown.ItemIndicator>
                         {({ isSelected }) => (
                           <AnimatePresence>
-                            {isSelected && <AnimatedCheck className='text-muted' stroke={1.5} />}
+                            {isSelected && <AnimatedCheck className='text-muted' />}
                           </AnimatePresence>
                         )}
                       </Dropdown.ItemIndicator>
@@ -987,7 +987,7 @@ export function ActivityLogTable() {
                       <Dropdown.ItemIndicator>
                         {({ isSelected }) => (
                           <AnimatePresence>
-                            {isSelected && <AnimatedCheck className='text-muted' stroke={1.5} />}
+                            {isSelected && <AnimatedCheck className='text-muted' />}
                           </AnimatePresence>
                         )}
                       </Dropdown.ItemIndicator>