+++ src/components/DataflowInspector.jsx
@@ -374,7 +374,7 @@ export function DataflowInspector({
                   matchesURL={false}
                   src={rawJSON}
                   CopiedComponent={({ className, style }) => (
-                    <AnimatedCheck className={className + ' text-success'} size={16} stroke={1.5} style={style} />
+                    <AnimatedCheck className={className + ' text-success'} size={16} style={style} />
                   )}
                   CopyComponent={({ className, onClick, style }) => (
                     <IconClipboardCopy className={className} size={16} style={style} onClick={onClick} />