+++ src/components/GroupIdAnnotation.jsx
@@ -19,7 +19,7 @@ export function GroupIdAnnotation({ displayName, value }) {
     <>
       <span className={isString ? 'json-view--string' : 'json-view--number'}>{isString ? `"${value}"` : String(value)}</span>
       {copied ? (
-        <AnimatedCheck className='json-view--copy text-success' size={16} stroke={1.5} style={{ display: 'inline-block' }} />
+        <AnimatedCheck className='json-view--copy text-success' size={16} style={{ display: 'inline-block' }} />
       ) : (
         <IconClipboardCopy className='json-view--copy' size={16} onClick={handleCopy} />
       )}