+++ src/components/views/ApiErrorsView.jsx
@@ -168,7 +168,7 @@ export function ApiErrorsView({ instance = null, onBackToDefault = null, onStatu
                         matchesURL={false}
                         src={parsed}
                         CopiedComponent={({ className, style }) => (
-                          <AnimatedCheck className={className + ' text-success'} size={16} stroke={1.5} style={style} />
+                          <AnimatedCheck className={className + ' text-success'} size={16} style={style} />
                         )}
                         CopyComponent={({ className, onClick, style }) => (
                           <IconClipboardCopy className={className} size={16} style={style} onClick={onClick} />