+++ src/components/views/ObjectDetailsView.jsx
@@ -223,7 +223,7 @@ export function ObjectDetailsView({ instance = null, liveContext = null, onBackT
                     matchesURL={false}
                     src={domoObject.metadata?.details}
                     CopiedComponent={({ className, style }) => (
-                      <AnimatedCheck className={className} size={16} stroke={1.5} style={style} />
+                      <AnimatedCheck className={className} size={16} style={style} />
                     )}
                     CopyComponent={({ className, onClick, style }) => (
                       <IconClipboardCopy className={className} size={16} style={style} onClick={onClick} />